requirements.txt
pytest==3.3.1
i3ipc==1.3.0
setuptools==28.8.0
pypandoc==1.4
ruamel.base==1.0.0
